+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.24.3
 
 ### Bug Fix
-- Fixed inability to download exports
+- Fixed inability to download project exports
 - Fixed inability to download project documents
 
 ## 0.24.2 (April 18, 2014)